Move form element definitions
