+++ apps/doj-demo/astro.config.mjs
@@ -5,7 +5,7 @@ import react from '@astrojs/react';
 // https://astro.build/config
 export default defineConfig({
   output: 'server',
-  trailingSlash: 'never',
+  trailingSlash: 'always',
   base: addTrailingSlash(process.env.BASEURL || ''),
   adapter: node({
     mode: 'standalone',
