+++ packages/forms/src/elements/input.ts
@@ -1,7 +1,6 @@
 import * as z from 'zod';
 
-import { type FormElementConfig } from '..';
-import { type FormElement, validateElement } from '../element';
+import { type FormElement, type FormElementConfig } from '../element';
 import { type Pattern, type TextInputPattern } from '../pattern';
 import { getFormSessionValue } from '../session';
 import { safeZodParse } from '../util/zod';
